Work delete functions and delte image from database and server

+++ resources/views/welcome.blade.php
@@ -11,7 +11,7 @@
                         <div class="d-grid gap-2">
                             <a href="/show/{{$image->id}}" class="btn btn-info">Show</a>
                             <a href="/edit/{{$image->id}}" class="btn btn-warning">Edit</a>
-                            <a href="/delete/{{$image->id}}" class="btn btn-danger">Delete</a>
+                            <a href="/delete/{{$image->id}}" onclick="return confirm('Are you sure?')" class="btn btn-danger">Delete</a>
                         </div>
                     </div>
                 </div>
